TeacherReviewManualExam functionality has been added

+++ CommonLibs/src/main/java/DatabaseAccess/Requests/ExecuteExam/EvaluateManualExamRequest.java
@@ -5,17 +5,17 @@ import DatabaseAccess.Requests.DatabaseRequest;
 import java.io.File;
 
 public class EvaluateManualExamRequest extends DatabaseRequest {
-    private final int grade;
+    private final double grade;
     private final String comments;
     private final File wordFile;
 
-    public EvaluateManualExamRequest(int grade, String comments, File wordFile) {
+    public EvaluateManualExamRequest(double grade, String comments, File wordFile) {
         this.grade = grade;
         this.comments = comments;
         this.wordFile = wordFile;
     }
 
-    public int getGrade() {
+    public double getGrade() {
         return grade;
     }
